Declare a random generator namespace
Later on, we would like to be able to randomly place the apple on the screen for the snake to eat. For that to happen, I am going to define random generator classes for both int and float numbers. The float random generator will be used to generate random numbers needed for the coordinates of the position of the apple. The int one is going to be used in order to randomly determine the interval for which if we choose any point in it, no parts of the snake occupy that position.