Define enum class RelativePosition
We should be able to know the relative position of squares with respect to the square before them. This gets a step closer to being able to determine the movement of the specified square when the square that they are attached to moves in a certain direction. We are also using enum class due to type safety issues that normal enums have.