Update movement direction
We have designed our snake to be composed of several green squares. The movement of the snake is going to be dictated by the head square. When the head square moves, the most important thing is its movement direction. That direction is going to tell the closest attached square behind it as to which direction it should move. That attached sqaure is going to move and in the same way its direction is going to determine which direction the closest attached square to it is going to move. We shall iterate the same mechanism for all the attached squares until we reach the tail of the snake.

In order to implement this idea, I defined UpdateMovementDirection(). This function updates the movement direction data member so that we could then use this info and move accordingly. It has Set_RelativePos*MovementDir() functions as helper functions. They explicitely set the m_movementDirection data member and make  UpdateMovementDirection() more readable. Since the sole purpose of these helper functions is to help update movement direction, they are in the private scope of the GreenSquare class.